Revert EntityManagerUtils change
